+++ platform-sdk/swirlds-platform-core/src/main/java/module-info.java
@@ -61,6 +61,7 @@ module com.swirlds.platform {
     /* Targeted Exports to External Libraries */
     exports com.swirlds.platform.internal to
             com.swirlds.platform.test,
+            com.swirlds.platform.test.fixtures,
             com.fasterxml.jackson.core,
             com.fasterxml.jackson.databind;
     exports com.swirlds.platform.event.creation to
