+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -1,0 +1,8 @@
+open module com.swirlds.platform.test.fixtures {
+    requires com.swirlds.common;
+    requires com.swirlds.platform;
+    requires org.apache.commons.lang3;
+
+    exports com.swirlds.platform.test.fixtures.stream;
+    exports com.swirlds.platform.test.fixtures.event;
+}
