+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/consensus/ConsensusUtils.java
@@ -35,9 +35,9 @@ import com.swirlds.platform.metrics.ConsensusMetrics;
 import com.swirlds.platform.test.NoOpConsensusMetrics;
 import com.swirlds.platform.test.NoOpIntakeCycleStats;
 import com.swirlds.platform.test.event.EventUtils;
-import com.swirlds.platform.test.event.IndexedEvent;
 import com.swirlds.platform.test.event.TestSequence;
 import com.swirlds.platform.test.event.emitter.EventEmitter;
+import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
