+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/consensus/NodeContext.java
@@ -26,10 +26,10 @@ import com.swirlds.platform.eventhandling.SignedStateEventsAndGenerations;
 import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.state.signed.SignedState;
 import com.swirlds.platform.test.event.EventUtils;
-import com.swirlds.platform.test.event.IndexedEvent;
 import com.swirlds.platform.test.event.emitter.CollectingEventEmitter;
 import com.swirlds.platform.test.event.emitter.EventEmitter;
 import com.swirlds.platform.test.event.source.EventSource;
+import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Arrays;
