+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/event/DynamicValueGenerator.java
@@ -16,6 +16,7 @@
 
 package com.swirlds.platform.test.event;
 
+import com.swirlds.platform.test.fixtures.event.DynamicValue;
 import java.util.Random;
 
 /**
