+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/event/source/EventSource.java
@@ -20,8 +20,8 @@ import static com.swirlds.platform.test.event.EventUtils.staticDynamicValue;
 
 import com.swirlds.common.system.NodeId;
 import com.swirlds.common.test.fixtures.TransactionGenerator;
-import com.swirlds.platform.test.event.DynamicValue;
-import com.swirlds.platform.test.event.IndexedEvent;
+import com.swirlds.platform.test.fixtures.event.DynamicValue;
+import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.time.Instant;
 import java.util.Random;
