+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/event/source/StandardEventSource.java
@@ -17,7 +17,7 @@
 package com.swirlds.platform.test.event.source;
 
 import com.swirlds.common.test.fixtures.TransactionGenerator;
-import com.swirlds.platform.test.event.IndexedEvent;
+import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import java.util.LinkedList;
 import java.util.Random;
