+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/module-info.java
@@ -12,4 +12,5 @@ open module com.swirlds.platform.test {
     requires static com.github.spotbugs.annotations;
     requires com.swirlds.config.api.test.fixtures;
     requires com.swirlds.common.test.fixtures;
+    requires com.swirlds.platform.test.fixtures;
 }
