+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/cli/EventStreamSingleFileRepairTest.java
@@ -32,8 +32,8 @@ import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.recovery.internal.EventStreamSingleFileIterator;
 import com.swirlds.platform.recovery.internal.EventStreamSingleFileRepairer;
 import com.swirlds.platform.test.consensus.GenerateConsensus;
+import com.swirlds.platform.test.fixtures.stream.StreamUtils;
 import com.swirlds.platform.test.simulated.RandomSigner;
-import com.swirlds.platform.test.stream.StreamUtils;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
