+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/consensus/ConsensusTestDefinitions.java
@@ -54,9 +54,7 @@ import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.internal.SignedStateLoadingException;
 import com.swirlds.platform.state.signed.SignedState;
 import com.swirlds.platform.test.SimpleEventGenerator;
-import com.swirlds.platform.test.event.DynamicValue;
 import com.swirlds.platform.test.event.EventUtils;
-import com.swirlds.platform.test.event.IndexedEvent;
 import com.swirlds.platform.test.event.TestSequence;
 import com.swirlds.platform.test.event.emitter.PriorityEventEmitter;
 import com.swirlds.platform.test.event.emitter.StandardEventEmitter;
@@ -65,6 +63,8 @@ import com.swirlds.platform.test.event.source.EventSource;
 import com.swirlds.platform.test.event.source.EventSourceFactory;
 import com.swirlds.platform.test.event.source.ForkingEventSource;
 import com.swirlds.platform.test.event.source.StandardEventSource;
+import com.swirlds.platform.test.fixtures.event.DynamicValue;
+import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.ArrayList;
